Cleans up podcast creation view, creates custom permacast theme
