+++ src/utils/arweave.js
@@ -1,7 +1,7 @@
 import ArweaveMultihost from "arweave-multihost";
 import Arweave from "arweave";
 import { SmartWeaveWebFactory } from "redstone-smartweave";
-
+import { generateFactoryState } from "./initStateGen";
 /*
 export const arweave = ArweaveMultihost.initWithDefaultHosts({
   timeout: 10000,         // Network request timeouts in milliseconds
@@ -58,336 +58,23 @@ export const queryObject = {
   }`,
 };
 
-export const categories_en = [
-  "Arts",
-  "Business",
-  "Comedy",
-  "Education",
-  "Fiction",
-  "Government",
-  "History",
-  "Kids & Family",
-  "Leisure",
-  "Music",
-  "News",
-  "Religion & Spirituality",
-  "Science",
-  "Society & Culture",
-  "Sports",
-  "Technology",
-  "True Crime",
-  "TV & Film",
-];
+export async function deployContract (address) {
 
-export const categories_zh = [
-  "艺术",
-  "商业",
-  "喜剧",
-  "教育",
-  "小说",
-  "政府",
-  "历史",
-  "儿童 & 家庭",
-  "休闲",
-  "音乐",
-  "新闻",
-  "宗教 & 灵修",
-  "科学",
-  "社会 & 文化",
-  "体育",
-  "科技",
-  "真相",
-  "电视 & 电影",
-];
+  const initialState = await generateFactoryState(address);
+  console.log(initialState)
+  const tx = await arweave.createTransaction({ data: initialState })
 
-// https://meta.wikimedia.org/wiki/Template:List_of_language_names_ordered_by_code
-export const languages_en = {
-  en: "English",
-  aa: "Afar",
-  ab: "Abkhazian",
-  af: "Afrikaans",
-  am: "Amharic",
-  ar: "Arabic",
-  as: "Assamese",
-  ay: "Aymara",
-  az: "Azeri",
-  ba: "Bashkir",
-  be: "Belarusian",
-  bg: "Bulgarian",
-  bh: "Bihari",
-  bi: "Bislama",
-  bn: "Bengali",
-  bo: "Tibetan",
-  br: "Breton",
-  ca: "Catalan",
-  co: "Corsican",
-  cs: "Czech",
-  cy: "Welsh",
-  da: "Danish",
-  de: "German",
-  div: "Divehi",
-  dz: "Bhutani",
-  el: "Greek",
-  eo: "Esperanto",
-  es: "Spanish",
-  et: "Estonian",
-  eu: "Basque",
-  fa: "Farsi",
-  fi: "Finnish",
-  fj: "Fiji",
-  fo: "Faeroese",
-  fr: "French",
-  fy: "Frisian",
-  ga: "Irish",
-  gd: "Gaelic",
-  gl: "Galician",
-  gn: "Guarani",
-  gu: "Gujarati",
-  ha: "Hausa",
-  he: "Hebrew",
-  hi: "Hindi",
-  hr: "Croatian",
-  hu: "Hungarian",
-  hy: "Armenian",
-  ia: "Interlingua",
-  id: "Indonesian",
-  ie: "Interlingue",
-  ik: "Inupiak",
-  is: "Icelandic",
-  it: "Italian",
-  ja: "Japanese",
-  jw: "Javanese",
-  ka: "Georgian",
-  kk: "Kazakh",
-  kl: "Greenlandic",
-  km: "Cambodian",
-  kn: "Kannada",
-  ko: "Korean",
-  kok: "Konkani",
-  ks: "Kashmiri",
-  ku: "Kurdish",
-  ky: "Kirghiz",
-  kz: "Kyrgyz",
-  la: "Latin",
-  ln: "Lingala",
-  lo: "Laothian",
-  lt: "Lithuanian",
-  lv: "Latvian",
-  mg: "Malagasy",
-  mi: "Maori",
-  mk: "FYRO Macedonian",
-  ml: "Malayalam",
-  mn: "Mongolian",
-  mo: "Moldavian",
-  mr: "Marathi",
-  ms: "Malay",
-  mt: "Maltese",
-  my: "Burmese",
-  na: "Nauru",
-  ne: "Nepali (India)",
-  nl: "Dutch",
-  no: "Norwegian",
-  oc: "Occitan",
-  om: "(Afan)/Oromoor/Oriya",
-  or: "Oriya",
-  pa: "Punjabi",
-  pl: "Polish",
-  ps: "Pashto/Pushto",
-  pt: "Portuguese",
-  qu: "Quechua",
-  rm: "Rhaeto-Romanic",
-  rn: "Kirundi",
-  ro: "Romanian",
-  ru: "Russian",
-  rw: "Kinyarwanda",
-  sa: "Sanskrit",
-  sb: "Sorbian",
-  sd: "Sindhi",
-  sg: "Sangro",
-  sh: "Serbo-Croatian",
-  si: "Singhalese",
-  sk: "Slovak",
-  sl: "Slovenian",
-  sm: "Samoan",
-  sn: "Shona",
-  so: "Somali",
-  sq: "Albanian",
-  sr: "Serbian",
-  ss: "Siswati",
-  st: "Sesotho",
-  su: "Sundanese",
-  sv: "Swedish",
-  sw: "Swahili",
-  sx: "Sutu",
-  syr: "Syriac",
-  ta: "Tamil",
-  te: "Telugu",
-  tg: "Tajik",
-  th: "Thai",
-  ti: "Tigrinya",
-  tk: "Turkmen",
-  tl: "Tagalog",
-  tn: "Tswana",
-  to: "Tonga",
-  tr: "Turkish",
-  ts: "Tsonga",
-  tt: "Tatar",
-  tw: "Twi",
-  uk: "Ukrainian",
-  ur: "Urdu",
-  uz: "Uzbek",
-  vi: "Vietnamese",
-  vo: "Volapuk",
-  wo: "Wolof",
-  xh: "Xhosa",
-  yi: "Yiddish",
-  yo: "Yoruba",
-  zh: "Chinese",
-  zu: "Zulu",
-};
-
-export const languages_zh = {
-  en: "英语",
-  aa: "阿法尔语",
-  ab: "阿布哈西亚语",
-  af: "南非荷兰语",
-  am: "阿姆哈拉语",
-  ar: "阿拉伯语",
-  as: "阿萨姆语",
-  ay: "艾马里语",
-  az: "阿塞拜疆语",
-  ba: "巴什基尔语",
-  be: "白俄罗斯语",
-  bg: "保加利亚语",
-  bh: "比哈尔语",
-  bi: "比斯拉姆语",
-  bn: "孟加拉语",
-  bo: "藏语",
-  br: "布列塔尼语",
-  ca: "加泰罗尼亚语",
-  co: "科西嘉语",
-  cs: "捷克语",
-  cy: "威尔士语",
-  da: "丹麦语",
-  de: "德语",
-  div: "迪维希语",
-  dz: "不丹语",
-  el: "希腊语",
-  eo: "世界语",
-  es: "西班牙语",
-  et: "爱沙尼亚语",
-  eu: "巴斯克语",
-  fa: "波斯语",
-  fi: "芬兰语",
-  fj: "斐济语",
-  fo: "法罗语",
-  fr: "法语",
-  fy: "弗里西语",
-  ga: "爱尔兰语",
-  gd: "苏格兰盖尔语",
-  gl: "加利西亚语",
-  gn: "瓜拉尼语",
-  gu: "古吉拉特语",
-  ha: "豪萨语",
-  he: "希伯来语",
-  hi: "印地语",
-  ho: "希里莫图语",
-  hr: "克罗地亚语",
-  hu: "匈牙利语",
-  hy: "亚美尼亚语",
-  ia: "因为语",
-  id: "印度尼西亚语",
-  ie: "因纽特语",
-  ik: "依奴皮克语",
-  is: "冰岛语",
-  it: "意大利语",
-  ja: "日语",
-  jw: "爪哇语",
-  ka: "格鲁吉亚语",
-  kk: "哈萨克语",
-  kl: "格陵兰语",
-  km: "高棉语",
-  kn: "卡纳达语",
-  ko: "韩语",
-  ks: "克什米尔语",
-  ku: "库尔德语",
-  ky: "吉尔吉斯语",
-  la: "拉丁语",
-  lb: "卢森堡语",
-  lg: "广东语",
-  li: "林堡语",
-  ln: "林加拉语",
-  lo: "老挝语",
-  lt: "立陶宛语",
-  lv: "拉脱维亚语",
-  mg: "马尔加什语",
-  mi: "毛利语",
-  mk: "马其顿语",
-  ml: "马拉雅拉姆语",
-  mn: "蒙古语",
-  mo: "摩尔多瓦语",
-  mr: "马拉地语",
-  ms: "马来语",
-  mt: "马耳他语",
-  my: "缅甸语",
-  na: "瑙鲁语",
-  ne: "尼泊尔语",
-  nl: "荷兰语",
-  no: "挪威语",
-  oc: "奥克语",
-  om: "奥米语",
-  or: "奥利亚语",
-  pa: "旁遮普语",
-  pl: "波兰语",
-  ps: "普什图语",
-  pt: "葡萄牙语",
-  qu: "克丘亚语",
-  rm: "罗曼什语",
-  rn: "埃塞俄比亚语",
-  ro: "罗马尼亚语",
-  ru: "俄语",
-  rw: "卢旺达语",
-  sa: "梵文",
-  sd: "信德语",
-  sg: "桑戈语",
-  sh: "塞尔维亚-克罗地亚语",
-  si: "僧伽罗语",
-  sk: "斯洛伐克语",
-  sl: "斯洛文尼亚语",
-  sm: "萨摩亚语",
-  sn: "修纳语",
-  so: "索马里语",
-  sq: "阿尔巴尼亚语",
-  sr: "塞尔维亚语",
-  ss: "斯威士语",
-  st: "塞索托语",
-  su: "巽他语",
-  sv: "瑞典语",
-  sw: "斯瓦希里语",
-  ta: "泰米尔语",
-  te: "泰卢固语",
-  tg: "塔吉克语",
-  th: "泰语",
-  ti: "提格里尼亚语",
-  tk: "土库曼语",
-  tl: "菲律宾语",
-  tn: "茨瓦纳语",
-  to: "汤加语",
-  tr: "土耳其语",
-  ts: "宗加语",
-  tt: "塔塔尔语",
-  tw: "特威语",
-  ug: "维吾尔语",
-  uk: "乌克兰语",
-  ur: "乌尔都语",
-  uz: "乌兹别克语",
-  vi: "越南语",
-  vo: "沃拉普克语",
-  wo: "沃洛夫语",
-  xh: "卡索语",
-  yi: "依地语",
-  yo: "约鲁巴语",
-  za: "壮语",
-  zh: "中文",
-  zu: "祖鲁语",
-};
+  tx.addTag("App-Name", "SmartWeaveContract")
+  tx.addTag("App-Version", "0.3.0")
+  tx.addTag("Contract-Src", CONTRACT_SRC)
+  tx.addTag("Permacast-Version", "amber");
+  tx.addTag("Content-Type", "application/json")
+  tx.addTag("Timestamp", Date.now())
+  
+  tx.reward = (+tx.reward * FEE_MULTIPLIER).toString();
+  
+  await arweave.transactions.sign(tx)
+  await arweave.transactions.post(tx)
+  console.log(tx)
+  return tx.id
+}
