+++ tailwind.config.js
@@ -4,6 +4,17 @@ module.exports = {
   theme: {
     extend: {},
   },
+  daisyui: {
+    themes: [
+      "business",
+      {
+        permacast: {
+          primary: "#FFFF00",
+          secondary: "#18181b",
+        }
+      }
+    ]
+  },
   plugins: [
     require("@tailwindcss/aspect-ratio"),
     require("@tailwindcss/typography"),
